+++ sdk/samples/iot/pnp/pnp_mqtt_message.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "pnp_mqtt_component.h"
+#include "pnp_mqtt_message.h"
 
 #include <stddef.h>
 #include <stdint.h>
