+++ sdk/samples/iot/pnp/pnp_mqtt_message.h
